Made the /tmp/gps_direction.out reader more general.
It now accepts any number of values on each line, only acting on the last one
(space separated).

It also backs up 120 characters, then does a readline to get in sync with the
last line or two of the file when it's open.  It used to just jump to the end
of the file and wait for more data to be written to the file.  Now it will see
the last 1 or 2 values.

These changes let you manually enter data into the /tmp/gps_direction.out file
for testing.